+++ crates/matrix-sdk-base/CHANGELOG.md
@@ -17,6 +17,9 @@ All notable changes to this project will be documented in this file.
 - [**breaking**] `RoomInfoNotableUpdateReasons` is now a `u16` to include a `FULLY_READ` flag to notify on changes of the `m.fully_read` marker.
   ([#6569](https://github.com/matrix-org/matrix-rust-sdk/pull/6569))
 
+- Add `RoomInfo::fully_read_event_id` and `Room::fully_read_event_id` to expose the user's `m.fully_read` event ID.
+  ([#6569](https://github.com/matrix-org/matrix-rust-sdk/pull/6569))
+
 ## [0.17.0] - 2026-05-08
 
 ### Bug Fixes